Use float for heuristics calculation

+++ UnitTest/MockHeurCalculator.hpp
@@ -5,9 +5,9 @@
 #include <turtle/mock.hpp>
 
 MOCK_BASE_CLASS(MockHeurCalculator, HeurCalculator) {
-	MOCK_CONST_METHOD(calculateStatus, 3, int(
+	MOCK_CONST_METHOD(calculateStatus, 3, float(
 			const Status&, const MoveDescriptor*, const std::shared_ptr<Node>&))
-	MOCK_CONST_METHOD(calculateStone, 2, int(const Status&, Point))
+	MOCK_CONST_METHOD(calculateStone, 2, float(const Status&, Point))
 };
