+++ src/HeurCalculator.hpp
@@ -13,8 +13,8 @@ class HeurCalculator: public boost::noncopyable {
 public:
 	typedef std::shared_ptr<const HeurCalculator> Ptr;
 	virtual ~HeurCalculator() {}
-	virtual int calculateStone(const Status& status, Point  p) const = 0;
-	virtual int calculateStatus(
+	virtual float calculateStone(const Status& status, Point  p) const = 0;
+	virtual float calculateStatus(
 			const Status& status,
 			const MoveDescriptor* moveDescriptor = nullptr,
 			const std::shared_ptr<Node>& ancestor = std::shared_ptr<Node>()) const = 0;
